--- test/default.json
@@ -1,48 +1,0 @@
-{
-    "userId": 1,
-    "userName": "hoge",
-    "studyInfos": [
-	{
-	    "studyId": 1,
-	    "subId": 0,
-	    "studyTime": 1100,
-	    "dateTime": "2020-05-03 12:23:34"
-	},
-	{
-	    "studyId": 2,
-	    "subId": 0,
-	    "studyTime": 1234,
-	    "dateTime": "2020-05-04 01:02:03"
-	},
-	{
-	    "studyId": 5,
-	    "subId": 1,
-	    "studyTime": 11,
-	    "dateTime": "2020-05-05 12:23:34"
-	},
-	{
-	    "studyId": 7,
-	    "subId": 0,
-	    "studyTime": 300,
-	    "dateTime": "2020-05-06 12:23:34"
-	},
-	{
-	    "studyId": 8,
-	    "subId": 2,
-	    "studyTime": 4321,
-	    "dateTime": "2020-05-06 12:50:34"
-	},
-	{
-	    "studyId": 10,
-	    "subId": 0,
-	    "studyTime": 500,
-	    "dateTime": "2020-05-06 16:23:34"
-	},
-	{
-	    "studyId": 21,
-	    "subId": 1,
-	    "studyTime": 1100,
-	    "dateTime": "2020-05-09 12:23:34"
-	}
-    ]
-}
